minutes

+++ db/migrate/20231104064815_create_study_records.rb
@@ -3,7 +3,7 @@ class CreateStudyRecords < ActiveRecord::Migration[6.1]
     create_table :study_records do |t|
       t.string :title
       t.text :body
-      t.integer :num_of_session
+      t.float :num_of_session
       t.integer :user_id
       t.timestamps
     end
